requirements.txt
kubernetes
requests
PyYaml